+++ src/loki/locate_action.cc
@@ -22,30 +22,26 @@ std::string loki_worker_t::locate(Api& request) {
 
   // correlate the various locations to the underlying graph
   init_locate(request);
-  auto locations = PathLocation::fromPBF(request.options().locations());
-  const auto& costing_name = Costing_Enum_Name(request.options().costing_type());
-  std::unordered_map<baldr::Location, PathLocation> projections;
+  auto* options = request.mutable_options();
+  auto* locations = options->mutable_locations();
+  const auto& costing_name = Costing_Enum_Name(options->costing_type());
   if (costing_name == "auto_pedestrian") {
-    if (locations.size() > 2) {
-      throw valhalla_exception_t{150, "for auto_pedestrian: " + std::to_string(locations.size())};
+    if (locations->size() > 2) {
+      throw valhalla_exception_t{150, "for auto_pedestrian: " + std::to_string(locations->size())};
     }
-    std::vector<baldr::Location> start_loc(locations.begin(), locations.begin() + 1);
-    auto start_projection =
-        search_.search(start_loc, mode_costing[static_cast<size_t>(TravelMode::kDrive)]);
-    for (const auto& [loc, path_loc] : start_projection) {
-      projections.insert({loc, path_loc});
-    }
-    std::vector<baldr::Location> end_loc(locations.end() - 1, locations.end());
-
-    auto end_projection =
-        search_.search(end_loc, mode_costing[static_cast<size_t>(TravelMode::kPedestrian)]);
-    for (const auto& [loc, path_loc] : end_projection) {
-      projections.insert({loc, path_loc});
-    }
+    google::protobuf::RepeatedPtrField<Location> start_loc(locations->begin(),
+                                                           locations->begin() + 1);
+    search_.search(start_loc, mode_costing[static_cast<size_t>(mode)]);
+    google::protobuf::RepeatedPtrField<Location> end_loc(locations->begin() + 1,
+                                                         locations->begin() + 2);
+    search_.search(end_loc, mode_costing[static_cast<size_t>(mode)]);
+    // merge them again
+    locations->at(0).CopyFrom(start_loc.at(0));
+    locations->at(1).CopyFrom(end_loc.at(0));
   } else {
-    projections = search_.search(locations, mode_costing[static_cast<size_t>(mode)]);
+    search_.search(*locations, mode_costing[static_cast<size_t>(mode)]);
   }
-  return tyr::serializeLocate(request, locations, projections, *reader);
+  return tyr::serializeLocate(request, *reader);
 }
 
 } // namespace loki